backend: change directory structure for first ac update
