+++ atcoder-problems-backend/Cargo.toml
@@ -19,12 +19,11 @@ serde_json = "1.0"
 
 # SQL
 sql-client = { path = "./sql-client" }
+# SQL (for test)
+sqlx = { version = "0.5.1", features = ["postgres", "runtime-async-std-rustls"] }
 
 # Scraping
 atcoder-client = { path = "./atcoder-client" }
-
-# s3
-s3-client = { path = "./s3-client" }
 
 # Web framework
 tide = "0.16"
@@ -40,4 +39,4 @@ futures = "0.3.5"
 [dev-dependencies]
 
 [workspace]
-members = ["sql-client", "atcoder-client", "s3-client"]
+members = ["sql-client", "atcoder-client"]
