--- atcoder-problems-backend/s3-client/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "s3-client"
-version = "0.1.0"
-authors = ["kenkoooo <kenkou.n@gmail.com>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-log = "0.4"
-rust-s3 = "=0.18.6"
-anyhow = "1.0.32"
