--- atcoder-problems-backend/s3-client/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod s3;
