+++ atcoder-problems-backend/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod crawler;
 pub mod server;
+pub mod s3;
+pub mod updater;
 pub mod utils;
 pub mod config;
