+++ atcoder-problems-backend/src/updater/mod.rs
@@ -1,0 +1,1 @@
+pub mod first_ac_submission;
